+++ app/src/main/java/com/sopt/now/core/base/BindingFragment.kt
@@ -1,4 +1,4 @@
-package com.sopt.now.util.base
+package com.sopt.now.core.base
 
 import android.os.Bundle
 import android.view.LayoutInflater
