+++ app/src/main/java/com/sopt/now/core/base/HeaderInterceptor.kt
@@ -1,4 +1,4 @@
-package com.sopt.now.util
+package com.sopt.now.core.base
 
 import okhttp3.Interceptor
 import okhttp3.Response
