+++ app/src/main/java/com/sopt/now/presentation/search/SearchFragment.kt
@@ -2,7 +2,7 @@ package com.sopt.now.presentation.search
 
 import android.os.Bundle
 import android.view.View
-import com.sopt.now.util.base.BindingFragment
+import com.sopt.now.core.base.BindingFragment
 import com.sopt.now.databinding.FragmentSearchBinding
 
 class SearchFragment : BindingFragment<FragmentSearchBinding>(FragmentSearchBinding::inflate) {
